requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
drf-yasg>=1.21,<2.0
PyYAML>=6.0
django-cors-headers>=4.0